+++ src/app/services/api.ts
@@ -139,6 +139,7 @@ export class ApiService {
       'description',
       'details',
       'engagementStatus',
+      'logos',
       'backgroundInfo',
       'engagementInfo',
       'engagementLabel',