+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * Remove unused function params, libraries, variables, etc.
 * Remove old, commented-out code
 * Enabled image uploading through CKEditor
+* Add project logos
+* Add alternative text for documents
+* Add file upload modal
 
 ### 1.0.0: October 22, 2021
 * Updating version in package.json, and changed package name to landuseplanning-admin